pyproject.toml
[project]
name = "ctf-agent-two"
version = "0.1.0"
description = "CTF agent v2 — multi-model solver swarm"
requires-python = ">=3.14"
dependencies = [
    "pydantic-ai",
    "pydantic-settings",
    "genai-prices",
    "aiodocker",
    "claude-agent-sdk",
    "httpx",
    "pyyaml",
    "rich",
    "click",
    "markdownify",
    "playwright",
]

[project.scripts]
ctf-solve = "backend.cli:main"
ctf-import = "backend.import_cli:main"
ctf-bump = "backend.cli:bump"
ctf-msg = "backend.cli:msg"
ctf-status = "backend.cli:status"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3",
    "ruff>=0.14",
    "ty>=0.0.2",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
target-version = "py314"
line-length = 100
extend-exclude = [
    ".omx",
    ".shared-artifacts",
    ".venv",
    "2026_GMDSOFT",
    "logs",
    "**/.lane-state/**",
    "**/provider-home/**",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N", "C4", "PTH"]
ignore = ["E501", "SIM105", "SIM115", "PTH123"]

[tool.ruff.lint.isort]
known-first-party = ["backend"]

[tool.ty.rules]
unresolved-import = "ignore"

[tool.ty.environment]
python-version = "3.14"

[tool.ty.src]
exclude = [
    ".omx/**",
    ".shared-artifacts/**",
    ".venv/**",
    "2026_GMDSOFT/**",
    "logs/**",
    "**/.lane-state/**",
    "**/provider-home/**",
]
